apps/python/backend/pyproject.toml
[project]
name = "voice-flow-backend"
version = "0.1.0"
description = "Voice Flow Application."
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "livekit-api>=1.0.5",
    "phonenumbers>=9.0.13",
    "psycopg[c]>=3.2.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "voice-flow-shared",
]

[tool.uv.sources]
voice-flow-shared = { workspace = true }

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
